Fix warnings (#108)
* Fix warnings

* Fix warnings in submodules

* Hide npm warnings if the user_interface build is a success

* Fix duplicate tf, npm no longer clear terminal

* Fix install script for t-top v4 for ROS2, jetson orin and jetpack 5.1.4

* Update submodules

+++ ros/perceptions/ego_noise_reduction/src/ego_noise_reduction/WeightedAverageWithAPrioriNoiseEstimator.cpp
@@ -149,19 +149,19 @@ void WeightedAverageWithAPrioriNoiseEstimator::verifyParameters(
 
     string line;
     getline(file, line);
-    if (stoi(line) != frameSampleCount)
+    if (stoull(line) != frameSampleCount)
     {
         THROW_NOT_SUPPORTED_EXCEPTION("Invalid noise directory N_FFT.");
     }
 
     getline(file, line);
-    if (stoi(line) != samplingFrequency)
+    if (stoull(line) != samplingFrequency)
     {
         THROW_NOT_SUPPORTED_EXCEPTION("Invalid noise directory sampling frequency.");
     }
 
     getline(file, line);
-    if (stoi(line) != channelCount())
+    if (stoull(line) != channelCount())
     {
         THROW_NOT_SUPPORTED_EXCEPTION("Invalid noise directory channel count.");
     }
